+++ generators/widget/templates/react.activity.controller.jsx
@@ -1,9 +1,17 @@
 <%- banner %>
 import React from 'react';
+import { injections } from 'laxar-react-adapter';
 
-export const injections = [];
-export function create() {
+export default class <%= reactClassName %> extends React.Component {
+   static [ injections ] = [ 'axContext' ];
 
-   // TODO: add your implementation here
+   constructor( props ) {
+      super( props );
+      const [ context ] = props.injections;
+      // TODO: add your implementation here
+   }
 
+   render() {
+      return <div style={{display: 'none'}}/>;
+   }
 }